Update C code
* remove dplyr_group_keys_impl()
* move dplyr_group_indices() to group_data.cpp
