+++ src/group_by.cpp
@@ -313,59 +313,3 @@ SEXP dplyr_validate_grouped_df(SEXP df, SEXP s_nr_df, SEXP s_check_bounds) {
 
   return R_NilValue;
 }
-
-SEXP dplyr_group_keys_impl(SEXP data) {
-  SEXP keys;
-  SEXP keys_names;
-  SEXP groups = Rf_getAttrib(data, dplyr::symbols::groups);
-
-  R_xlen_t nr;
-
-  if (Rf_isNull(groups)) {
-    nr = 1;
-    keys_names = PROTECT(Rf_allocVector(STRSXP, 0));
-    keys = PROTECT(Rf_allocVector(VECSXP, 0));
-  } else {
-    R_xlen_t nc = XLENGTH(groups) - 1;
-    nr = XLENGTH(VECTOR_ELT(groups, nc));
-
-    SEXP groups_names = Rf_getAttrib(groups, R_NamesSymbol);
-
-    keys = PROTECT(Rf_allocVector(VECSXP, nc));
-    keys_names = PROTECT(Rf_allocVector(STRSXP, nc));
-    for (R_xlen_t i = 0; i < nc; i++) {
-      SET_VECTOR_ELT(keys, i, VECTOR_ELT(groups, i));
-      SET_STRING_ELT(keys_names, i, STRING_ELT(groups_names, i));
-    }
-  }
-
-  SEXP rn = PROTECT(Rf_allocVector(INTSXP, 2));
-  INTEGER(rn)[0] = NA_INTEGER;
-  INTEGER(rn)[1] = -nr;
-  Rf_setAttrib(keys, R_RowNamesSymbol, rn);
-
-  Rf_setAttrib(keys, R_NamesSymbol, keys_names);
-  Rf_setAttrib(keys, R_ClassSymbol, dplyr::vectors::classes_tbl_df);
-
-  UNPROTECT(3);
-  return keys;
-}
-
-SEXP dplyr_group_indices(SEXP rows, SEXP s_nr) {
-  R_xlen_t nr = INTEGER(s_nr)[0];
-  R_xlen_t ng = XLENGTH(rows);
-
-  SEXP indices = PROTECT(Rf_allocVector(INTSXP, nr));
-  int* p_indices = INTEGER(indices);
-  for (R_xlen_t i = 0; i < ng; i++) {
-    SEXP rows_i = VECTOR_ELT(rows, i);
-    R_xlen_t n_i = XLENGTH(rows_i);
-    int* p_rows_i = INTEGER(rows_i);
-    for (R_xlen_t j = 0; j < n_i; j++, ++p_rows_i) {
-      p_indices[*p_rows_i - 1] = i + 1;
-    }
-  }
-
-  UNPROTECT(1);
-  return indices;
-}
